backend-flask/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
PyJWT==2.9.0
gunicorn==22.0.0
twilio==9.7.0
redis==5.0.1
rq==2.5.0
flask-limiter==3.12
marshmallow==4.0.0

# Optional monitoring (add SENTRY_DSN env var to enable)
sentry-sdk[flask]==1.40.0
